% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{as.matrix_or_sparse}
\alias{as.matrix_or_sparse}
\title{Coerce expression data to a matrix or a sparse matrix}
\usage{
as.matrix_or_sparse(x)
}
\description{
Coerce expression data to a matrix or a sparse matrix
}
\keyword{internal}
