% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{collect_models_result}
\alias{collect_models_result}
\title{Turn per-type fit results into the weights table and (optionally) the
coefficient list of the model bundle}
\usage{
collect_models_result(res, uctypes, return_models)
}
\description{
Turn per-type fit results into the weights table and (optionally) the
coefficient list of the model bundle
}
\keyword{internal}
