% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{do_call}
\alias{do_call}
\title{Faster version of do.call}
\source{
Gmisc
}
\usage{
do_call(what, args, quote = FALSE, envir = parent.frame())
}
\arguments{
\item{what}{The function to call}

\item{args}{The arguments to pass to the function}

\item{quote}{Whether to quote the arguments}

\item{envir}{The environment to look for the function in the parent frame}
}
\value{
The same results as using do.call()
}
\description{
Faster version of do.call
}
\author{
Max Gordon \if{html}{\out{<max at gforge.se>}}
}
