% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{explode}
\alias{explode}
\title{Split the string by the separator and trim the whitespaces.}
\usage{
explode(x, sep = ",")
}
\arguments{
\item{x}{A string.}

\item{sep}{The separator.}
}
\value{
A character vector.
}
\description{
Split the string by the separator and trim the whitespaces.
}
