% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{find_markers_fc}
\alias{find_markers_fc}
\title{Find markers by fold change}
\usage{
find_markers_fc(
  exprs,
  clusters,
  top,
  min_log2fc,
  min_pct,
  pos_only,
  refs,
  max_pct_out
)
}
\description{
Find markers by fold change
}
\keyword{internal}
