% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{fold_models_lr}
\alias{fold_models_lr}
\title{Average per-fold linear predictor coefficients (cv_folds > 1)}
\usage{
fold_models_lr(fold_fits, return_models)
}
\description{
Average per-fold linear predictor coefficients (cv_folds > 1)
}
\keyword{internal}
