% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitype_score.R
\name{hitype_score}
\alias{hitype_score}
\title{Calculate cell type scores}
\usage{
hitype_score(exprs, gs, scaled = FALSE, norm = "sqrt", use_sensitivity = TRUE)
}
\arguments{
\item{exprs}{Input scRNA-seq data matrix
(rownames - gene names, colnames - cell names)}

\item{gs}{The gene sets prepared by \code{\link{gs_prepare}}
The \code{gene_sets} is used. One could also pass \code{gs$gene_sets} directly.}

\item{scaled}{Whether the input data is scaled or not}

\item{norm}{Normalization method for scoring. \code{"sqrt"} (default)
divides by \code{sqrt(n)}; \code{"weight"} divides by
\code{sum(abs(weights))}.}

\item{use_sensitivity}{Whether to multiply expression z-scores by
marker sensitivity scores. Set to \code{FALSE} when using learned
weights to avoid double-penalizing shared markers. Default is \code{TRUE}.}
}
\value{
A list of matrices of cell type scores for each level
(rownames - cell types, colnames - cell names)
}
\description{
Calculate cell type scores
}
\author{
Matt Mulvahill, Panwen Wang
}
