% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitypedb_full}
\alias{hitypedb_full}
\title{Gene sets from ScType (full version)}
\format{
A data frame with columns commonly used by \code{\link{gs_prepare}}:
\code{cellName}, \code{tissueType}, \code{geneSymbolmore1},
\code{geneSymbolmore2} and \code{level}.
}
\source{
\url{https://raw.githubusercontent.com/IanevskiAleksandr/sc-type/master/ScTypeDB_full.xlsx}
}
\usage{
hitypedb_full
}
\description{
The full marker database from
\href{https://github.com/IanevskiAleksandr/sc-type}{ScType}, bundled for
convenience. It can be passed directly to \code{\link{gs_prepare}}, e.g.
\code{gs_prepare(hitypedb_full, tissue_type = "Immune system")}.
}
\keyword{datasets}
