% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitypedb_short}
\alias{hitypedb_short}
\title{Gene sets from ScType (short version)}
\format{
A data frame with columns commonly used by \code{\link{gs_prepare}}:
\code{cellName}, \code{tissueType}, \code{geneSymbolmore1},
\code{geneSymbolmore2} and \code{level}.
}
\source{
\url{https://raw.githubusercontent.com/IanevskiAleksandr/sc-type/master/ScTypeDB_short.xlsx}
}
\usage{
hitypedb_short
}
\description{
The short marker database from
\href{https://github.com/IanevskiAleksandr/sc-type}{ScType}, bundled for
convenience. It can be passed directly to \code{\link{gs_prepare}}:
}
\details{
\code{gs <- gs_prepare(hitypedb_short, tissue_type = "Immune system")}
}
\keyword{datasets}
