% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{n_ending}
\alias{n_ending}
\title{Get the number of repetitive given ending characters.}
\usage{
n_ending(x, ending)
}
\arguments{
\item{x}{A string.}

\item{ending}{The ending character.}
}
\value{
The number of ending character.
}
\description{
Get the number of repetitive given ending characters.
}
\keyword{internal}
