% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_prepare.R
\name{parse_next_imm_levels}
\alias{parse_next_imm_levels}
\title{Parse next immediate levels of cell types}
\usage{
parse_next_imm_levels(nl_marks, all_next_levels)
}
\arguments{
\item{nl_marks}{The marks for the next levels of cell types}

\item{all_next_levels}{All possible next levels of cell types}
}
\value{
A vector of possible next immediate levels of cell types
}
\description{
Parse next immediate levels of cell types
}
