% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_prepare.R
\name{parse_next_levels}
\alias{parse_next_levels}
\title{Parse all next levels of current cell name}
\usage{
parse_next_levels(
  cell_name,
  level,
  max_level,
  nl_marks,
  cell_markers,
  next_imm_levels
)
}
\arguments{
\item{cell_name}{The current cell name}

\item{level}{The current level}

\item{max_level}{The max level}

\item{nl_marks}{The next level marks}

\item{cell_markers}{The cell markers data frame}

\item{next_imm_levels}{The next immediate levels}
}
\value{
A list of all possible cell names
}
\description{
Parse all next levels of current cell name
}
