% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{revert_cell_name}
\alias{revert_cell_name}
\title{Remove the suffix of the cell name.}
\usage{
revert_cell_name(cell_name)
}
\arguments{
\item{cell_name}{A cell name.}
}
\value{
The cell name without suffix.
}
\description{
Remove the suffix of the cell name.
}
\keyword{internal}
