% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{train_glmnet}
\alias{train_glmnet}
\title{Sparse logistic regression weights (glmnet)}
\usage{
train_glmnet(data, clusters, cv_folds = 5, return_models = FALSE)
}
\description{
Sparse logistic regression weights (glmnet)
}
\keyword{internal}
