% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{train_lr}
\alias{train_lr}
\title{Logistic regression weights (one-vs-rest)}
\usage{
train_lr(data, clusters, cv_folds = 1, return_models = FALSE)
}
\description{
Logistic regression weights (one-vs-rest)
}
\keyword{internal}
