% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{train_lrp}
\alias{train_lrp}
\title{LRP-based weights (neural network)}
\usage{
train_lrp(data, clusters, uclusters, cv_folds, epochs, batch_size, data_split)
}
\description{
LRP-based weights (neural network)
}
\keyword{internal}
