% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{train_lrp_single}
\alias{train_lrp_single}
\title{Train a single LRP model (helper)}
\usage{
train_lrp_single(
  z,
  clusters,
  uclusters,
  train_idx,
  test_idx,
  epochs,
  batch_size,
  data_split
)
}
\description{
Train a single LRP model (helper)
}
\keyword{internal}
