% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_prepare.R
\docType{data}
\name{UNIVERSAL_COL_ALIASES}
\alias{UNIVERSAL_COL_ALIASES}
\title{Canonical column names of the universal marker format}
\format{
An object of class \code{list} of length 8.
}
\usage{
UNIVERSAL_COL_ALIASES
}
\description{
Canonical names win over aliases. The hitype/ScType db columns
(cellName, geneSymbolmore1/2, tissueType...) never collide with these.
}
\keyword{internal}
