% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_print.R
\name{valid_cell_type}
\alias{valid_cell_type}
\title{Get the valid cell type for a given cell type}
\usage{
valid_cell_type(cell_type, gs)
}
\arguments{
\item{cell_type}{A vector cell types, with length the same as the number of
cell type levels}

\item{gs}{A list of cell names for each cell type level}
}
\value{
A string of valid cell types or NULL if the cell type is invalid
}
\description{
Get the valid cell type for a given cell type
}
