% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{compile_weights}
\alias{compile_weights}
\title{Compile the weights}
\usage{
compile_weights(weights, gs, level, range)
}
\arguments{
\item{weights}{A data frame with the weights}

\item{gs}{The gene sets}

\item{level}{The level of the gene sets}

\item{range}{The range of the weights}
}
\value{
A data frame as the db
}
\description{
Compile the weights
}
\keyword{internal}
