% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{find_markers_presto}
\alias{find_markers_presto}
\title{Find markers with presto wilcoxauc}
\usage{
find_markers_presto(
  exprs,
  clusters,
  top,
  min_log2fc,
  min_pct,
  only_pos,
  include_negative
)
}
\description{
Find markers with presto wilcoxauc
}
\keyword{internal}
