% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{find_markers_seurat}
\alias{find_markers_seurat}
\title{Find markers with Seurat FindAllMarkers}
\usage{
find_markers_seurat(
  exprs,
  clusters,
  top,
  min_log2fc,
  min_pct,
  only_pos,
  include_negative
)
}
\description{
Find markers with Seurat FindAllMarkers
}
\keyword{internal}
