% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_prepare.R
\name{gs_prepare}
\alias{gs_prepare}
\title{Prepare gene sets for hitype}
\usage{
gs_prepare(path_to_db_file, tissue_type = NULL)
}
\arguments{
\item{path_to_db_file}{A data frame with markers or
Path to the marker gene database file, it should be
a tab-delimited text or excel file with the following columns:
\itemize{
\item \code{tissueType}: The tissue type of the cell type to be annotated.
This column is required only if \code{cell_type} is specified.
\item \code{cellName}: The name of the cell type to be annotated.
\item \code{nextLevels}: Possible next levels of the cell type to be annotated.
Introduced by \code{hitype}, so that we can work with hierarchical cell.
Levels are separated by \verb{;}. Cell names at each level are separated
by \verb{,}. An exclamatory mark \code{!} at the beginning of a level means
that the cell names at this level are mutually exclusive. If the
levels are less than possible next levels, then the remaining levels
are all possible next levels. See the example below.
\item \code{geneSymbolmore1}: The gene symbols of the marker genes that are
expected to be expressed in the cell type to be annotated.
The genes can be suffixed with one or more \code{+}. More \code{+} means
higher expression level. For example, \verb{CD3E++} means the gene
\code{CD3E} is expected to be highly expressed in the cell type.
\item \code{geneSymbolmore2}: The gene symbols of the marker genes that are
expected not to be expressed in the cell type to be annotated.
\item \code{level}: The levels of the cell names. Introduced by \code{hitype}, so that
we can work with hierarchical cell names. Different levels of
\code{cellName}s are predicted separately. For example, If we have \code{CD4}
as level 1 and \code{Naive} as level 2, then our prediction for a cell type
could be \verb{CD4 Naive}.
The levels should start from 1 and be consecutive. #'
}}

\item{tissue_type}{The tissue type of the cell type to be annotated.
This requires the \code{tissueType} column in the marker gene database file.
If \code{tissue_type} is specified, then only the cell types in the specified
tissue type will be used for annotation. If \code{tissue_type} is not specified,
then all cell types in the marker gene database file will be used for
annotation.}
}
\value{
A list with gene_sets and next_levels. The structure looks like:
list(
gene_sets = list(
# level 1
list( CD4 = list(markers = c(...), weights = c(...)), ... ),
# level 2
list( Naive = list(markers = c(...), weights = c(...)), ... )
),
# All possible final cell names
cell_names = list(CD4 = list(Naive = c("Activated", "Proliferating")))
)
}
\description{
GNU General Public License v3.0
}
\examples{
# nextLevels example
# If we have the following cell types:
#   level  cellName  nextLevels
#   1      CD4       Naive,Memory
#   2      Naive
#   2      Memory
#   3      Activated
#   3      Proliferating
# Then possible final cell names are:
#   CD4 Naive Activated
#   CD4 Naive Proliferating
#   CD4 Memory Activated
#   CD4 Memory Proliferating
#
# If the `nextLevels` of CD4 is `!Naive`, then possible final cell names are:
#   CD4 Memory Activated
#   CD4 Memory Proliferating
#
# If the `nextLevels` of CD4 is `!`, then possible final cell names are:
#   CD4 Activated
#   CD4 Proliferating
#
# If the `nextLevels` of CD4 is `!Naive;!`, then possible final cell names
# are:
#   CD4 Memory

}
\author{
Matt Mulvahill, Panwen Wang
}
