% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitype_score.R
\name{hitype_assign}
\alias{hitype_assign}
\title{Generate scores for cell types for each level}
\usage{
hitype_assign(
  clusters,
  scores,
  gs = NULL,
  fallback = "Unknown",
  threshold = 0.05,
  top = 10
)
}
\arguments{
\item{clusters}{A named vector of original cluster assignments
(names - cell names, values - cluster assignments)}

\item{scores}{A list of matrices of cell type scores for each level}

\item{gs}{The gene sets prepared by \code{\link{gs_prepare}}
The \code{cell_names} is actually used. One could also pass \code{gs$cell_names}
directly.}

\item{fallback}{A fallback cell type if no cell type is assigned}

\item{threshold}{A threshold for low confidence cell type assignment
The cell types are only assigned for cells with scores higher than the
threshold * ncells.
(0 - 1, default 0.05)}

\item{top}{The number of top cell types to assign for each cluster in the
result.}
}
\value{
A dataframe with columns: \code{Level}, \code{Cluster}, \code{CellType} and \code{Score}.
For each level and cluster, the top cell types are returned.
You can use \code{\link{summary.hitype_result}} to print the combination of
cell types for each cluster.
}
\description{
Generate scores for cell types for each level
}
