% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitype_score.R
\name{hitype_score}
\alias{hitype_score}
\title{Calculate ScType scores and assign cell types}
\usage{
hitype_score(exprs, gs, scaled = FALSE)
}
\arguments{
\item{exprs}{Input scRNA-seq data matrix
(rownames - gene names, colnames - cell names)}

\item{gs}{The gene sets prepared by \code{\link{gs_prepare}}
The \code{gene_sets} is used. One could also pass \code{gs$gene_sets} directly.}

\item{scaled}{Whether the input data is scaled or not}
}
\value{
A list of matrices of cell type scores for each level
(rownames - cell types, colnames - cell names)
}
\description{
GNU General Public License v3.0
}
\author{
Matt Mulvahill, Panwen Wang
}
