% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{prepare_data_for_training}
\alias{prepare_data_for_training}
\title{Prepare the data for weight training}
\usage{
prepare_data_for_training(
  path_to_gs,
  exprs,
  level = 1,
  scaled = FALSE,
  clusters = NULL
)
}
\arguments{
\item{path_to_gs}{Path to the gene set file without weights}

\item{exprs}{The expression matrix, or a seurat object
(rows: genes, columns: samples/cells)}

\item{level}{The level of the gene sets to train weights for if
you have multiple levels of gene sets.}

\item{scaled}{Whether the expression matrix is scaled}

\item{clusters}{A named vector of cluster ids
If \code{exprs} is a seurat object, this is ignored. The cluster ids are
taken from the seurat object.}
}
\value{
A list with the gene sets, the z matrix and the clusters
}
\description{
Prepare the data for weight training
}
