% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_ext.R
\name{RunHitype}
\alias{RunHitype}
\alias{RunHitype.default}
\alias{RunHitype.Seurat}
\title{Run hitype_assign for a Seurat object}
\usage{
RunHitype(object, ...)

\method{RunHitype}{default}(object, ...)

\method{RunHitype}{Seurat}(
  object,
  gs = NULL,
  fallback = "Unknown",
  threshold = 0.05,
  level_weights = function(l) 1/(10^(l - 1)),
  make_unique = FALSE,
  slot = "data",
  assay = NULL,
  scaled = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{gs}{The gene list prepared by \code{\link{gs_prepare}}}

\item{fallback}{A fallback cell type if no cell type is assigned}

\item{threshold}{A threshold for low confidence cell type assignment
The cell types are only assigned for cells with scores higher than the
threshold.
(0 - 1, default 0.05)}

\item{level_weights}{The weights for each level of the hierarchy to calculate
the final cell type score
It should be either a numeric vector of length equal to the number of levels
or a single numeric value to be used for all levels
It can also be a function that takes the levels as input and returns a
numeric vectors as the weights.}

\item{make_unique}{Whether to make the cell type names unique}

\item{slot}{The slot to use for \code{GetAssayData}}

\item{assay}{The assay to use for \code{GetAssayData}}

\item{scaled}{Whether the data from \code{GetAssayData} is scaled}
}
\value{
The Seurat object with the cell types (named \code{hitype}) added to the
metadata
}
\description{
Run hitype_assign for a Seurat object
}
