% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_print.R
\name{summary.hitype_result}
\alias{summary.hitype_result}
\title{Summarize the hitype_result object}
\usage{
\method{summary}{hitype_result}(hitype_res, top = 1, level_weights = function(l) 1/(10^(l - 1)))
}
\arguments{
\item{top}{The top assigned cell types to return for each cluster}

\item{level_weights}{The weights for each level of the hierarchy to calculate
the final cell type score
It should be either a numeric vector of length equal to the number of levels
or a single numeric value to be used for all levels
It can also be a function that takes the levels as input and returns a
numeric vectors as the weights.}
}
\value{
The summary of the hitype_result object
}
\description{
Summarize the hitype_result object
}
