% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_weights.R
\name{train_weights}
\alias{train_weights}
\title{Train weights for the markers}
\usage{
train_weights(
  path_to_gs,
  exprs,
  level = 1,
  scaled = FALSE,
  clusters = NULL,
  data_split = c(0.7, 0.2, 0.1),
  epochs = 20,
  batch_size = 32
)
}
\arguments{
\item{path_to_gs}{Path to the gene set file without weights}

\item{exprs}{The expression matrix, or a seurat object
(rows: genes, columns: samples/cells)}

\item{level}{The level of the gene sets to train weights for if
you have multiple levels of gene sets.}

\item{scaled}{Whether the expression matrix is scaled}

\item{clusters}{A named vector of cluster ids
If \code{exprs} is a seurat object, this is ignored. The cluster ids are
taken from the seurat object.}

\item{data_split}{A vector of fractions for training, validation and
testing. If only two fractions are provided, no testing set will be used.}

\item{epochs}{The number of epochs to train}

\item{batch_size}{The batch size}
}
\value{
A data frame with the weights, that can be used directly by
gs_prepare.
}
\description{
Train weights for the markers
}
